Finished integrating and labelling Bank 2
This bank was missed by SMSExamine so I had to relabel all 16K of it by
hand. The author of Emulicious contacted me and provided the means with
his emulator to disassemble the bank properly, but it was too late
already. Thanks goes to him and I only wish he had contacted me earlier
:)