Fill in the rest of the ROM from an original
We only have source code for the first 64KB. For now, the rest of the ROM file (256KB) will be filled in from an original copy.

A bug remains with WLA-DX, which is calculating the ROM-size in the SEGA header incorrectly. Once fixed, this source will produce a 1:1 original ROM.